+++ crates/libs/bindgen/rustfmt.toml
@@ -1,0 +1,1 @@
+newline_style = "Unix"
