+++ crates/libs/bindgen/src/enums.rs
@@ -14,7 +14,11 @@ pub fn gen(gen: &Gen, def: TypeDef) -> TokenStream {
         .reader
         .type_def_fields(def)
         .filter_map(|field| {
-            if gen.reader.field_flags(field).contains(FieldAttributes::LITERAL) {
+            if gen
+                .reader
+                .field_flags(field)
+                .contains(FieldAttributes::LITERAL)
+            {
                 let field_name = to_ident(gen.reader.field_name(field));
                 let constant = gen.reader.field_constant(field).unwrap();
                 let value = gen.value(&gen.reader.constant_value(constant));
@@ -176,8 +180,13 @@ pub fn gen(gen: &Gen, def: TypeDef) -> TokenStream {
             });
         }
 
-        if gen.reader.type_def_flags(def).contains(TypeAttributes::WINRT) {
-            let signature = Literal::byte_string(gen.reader.type_def_signature(def, &[]).as_bytes());
+        if gen
+            .reader
+            .type_def_flags(def)
+            .contains(TypeAttributes::WINRT)
+        {
+            let signature =
+                Literal::byte_string(gen.reader.type_def_signature(def, &[]).as_bytes());
 
             tokens.combine(&quote! {
                 #features
@@ -190,8 +199,6 @@ pub fn gen(gen: &Gen, def: TypeDef) -> TokenStream {
                 }
             });
         }
-
-        tokens.combine(&extensions::gen(type_name));
     }
 
     tokens
