+++ crates/libs/bindgen/src/gen.rs
@@ -11,7 +11,14 @@ pub struct Gen<'a> {
 
 impl<'a> Gen<'a> {
     pub fn new(reader: &'a Reader) -> Self {
-        Self { reader, namespace: "", sys: false, cfg: false, doc: false, component: false }
+        Self {
+            reader,
+            namespace: "",
+            sys: false,
+            cfg: false,
+            doc: false,
+            component: false,
+        }
     }
 
     //
@@ -181,7 +188,11 @@ impl<'a> Gen<'a> {
                     }
                 }
                 TypeKind::Delegate => {
-                    if self.reader.type_def_flags(*def).contains(TypeAttributes::WINRT) {
+                    if self
+                        .reader
+                        .type_def_flags(*def)
+                        .contains(TypeAttributes::WINRT)
+                    {
                         quote! { *mut ::core::ffi::c_void }
                     } else {
                         self.type_def_name(*def, &[])
@@ -191,12 +202,20 @@ impl<'a> Gen<'a> {
             },
             Type::MutPtr((kind, pointers)) => {
                 let pointers_tokens = gen_mut_ptrs(*pointers);
-                let kind = if *pointers > 1 { self.type_name(kind) } else { self.type_abi_name(kind) };
+                let kind = if *pointers > 1 {
+                    self.type_name(kind)
+                } else {
+                    self.type_abi_name(kind)
+                };
                 quote! { #pointers_tokens #kind }
             }
             Type::ConstPtr((kind, pointers)) => {
                 let pointers_tokens = gen_const_ptrs(*pointers);
-                let kind = if *pointers > 1 { self.type_name(kind) } else { self.type_abi_name(kind) };
+                let kind = if *pointers > 1 {
+                    self.type_name(kind)
+                } else {
+                    self.type_abi_name(kind)
+                };
                 quote! { #pointers_tokens #kind }
             }
             Type::WinrtArray(kind) => self.type_abi_name(kind),
@@ -243,8 +262,14 @@ impl<'a> Gen<'a> {
         tokens
     }
     /// The signature params which are generic (along with their relative index)
-    pub fn generic_params<'b>(&'b self, params: &'b [SignatureParam]) -> impl Iterator<Item = (usize, &SignatureParam)> + 'b {
-        params.iter().filter(move |param| self.reader.signature_param_is_convertible(param)).enumerate()
+    pub fn generic_params<'b>(
+        &'b self,
+        params: &'b [SignatureParam],
+    ) -> impl Iterator<Item = (usize, &SignatureParam)> + 'b {
+        params
+            .iter()
+            .filter(move |param| self.reader.signature_param_is_convertible(param))
+            .enumerate()
     }
     /// The generic param names (i.e., `T` in `fn foo<T>()`)
     pub fn constraint_generics(&self, params: &[SignatureParam]) -> TokenStream {
@@ -294,7 +319,9 @@ impl<'a> Gen<'a> {
                 SignatureParamKind::IntoParam => {
                     let name: TokenStream = gen_name(position);
                     let into = self.type_name(&param.ty);
-                    tokens.combine(&quote! { #name: ::std::convert::Into<::windows::core::InParam<#into>>, });
+                    tokens.combine(
+                        &quote! { #name: ::std::convert::Into<::windows::core::InParam<#into>>, },
+                    );
                 }
                 SignatureParamKind::Into => {
                     let name: TokenStream = gen_name(position);
@@ -411,7 +438,8 @@ impl<'a> Gen<'a> {
         if namespace == self.namespace {
             quote! {}
         } else {
-            let is_external = namespace.starts_with("Windows.") && !self.namespace.starts_with("Windows");
+            let is_external =
+                namespace.starts_with("Windows.") && !self.namespace.starts_with("Windows");
             let mut relative = self.namespace.split('.').peekable();
             let mut namespace = namespace.split('.').peekable();
 
@@ -510,7 +538,15 @@ impl<'a> Gen<'a> {
         let guid = self.type_name(&Type::GUID);
         format!("{}::from_u128(0x{:08x?}_{:04x?}_{:04x?}_{:02x?}{:02x?}_{:02x?}{:02x?}{:02x?}{:02x?}{:02x?}{:02x?})", guid.into_string(), value.0, value.1, value.2, value.3, value.4, value.5, value.6, value.7, value.8, value.9, value.10).into()
     }
-    pub fn interface_core_traits(&self, def: TypeDef, _generics: &[Type], ident: &TokenStream, constraints: &TokenStream, phantoms: &TokenStream, features: &TokenStream) -> TokenStream {
+    pub fn interface_core_traits(
+        &self,
+        def: TypeDef,
+        _generics: &[Type],
+        ident: &TokenStream,
+        constraints: &TokenStream,
+        phantoms: &TokenStream,
+        features: &TokenStream,
+    ) -> TokenStream {
         let name = trim_tick(self.reader.type_def_name(def));
         quote! {
             #features
@@ -535,7 +571,13 @@ impl<'a> Gen<'a> {
             }
         }
     }
-    pub fn agile(&self, def: TypeDef, ident: &TokenStream, constraints: &TokenStream, features: &TokenStream) -> TokenStream {
+    pub fn agile(
+        &self,
+        def: TypeDef,
+        ident: &TokenStream,
+        constraints: &TokenStream,
+        features: &TokenStream,
+    ) -> TokenStream {
         if self.reader.type_def_is_agile(def) {
             quote! {
                 #features
@@ -547,7 +589,15 @@ impl<'a> Gen<'a> {
             quote! {}
         }
     }
-    pub fn async_get(&self, def: TypeDef, generics: &[Type], ident: &TokenStream, constraints: &TokenStream, _phantoms: &TokenStream, features: &TokenStream) -> TokenStream {
+    pub fn async_get(
+        &self,
+        def: TypeDef,
+        generics: &[Type],
+        ident: &TokenStream,
+        constraints: &TokenStream,
+        _phantoms: &TokenStream,
+        features: &TokenStream,
+    ) -> TokenStream {
         let mut kind = self.reader.type_def_async_kind(def);
         let mut async_generics = generics.to_vec();
 
@@ -567,7 +617,9 @@ impl<'a> Gen<'a> {
             quote! {}
         } else {
             let return_type = match kind {
-                AsyncKind::Operation | AsyncKind::OperationWithProgress => self.type_name(&async_generics[0]),
+                AsyncKind::Operation | AsyncKind::OperationWithProgress => {
+                    self.type_name(&async_generics[0])
+                }
                 _ => quote! { () },
             };
 
@@ -575,7 +627,9 @@ impl<'a> Gen<'a> {
                 AsyncKind::Action => quote! { AsyncActionCompletedHandler },
                 AsyncKind::ActionWithProgress => quote! { AsyncActionWithProgressCompletedHandler },
                 AsyncKind::Operation => quote! { AsyncOperationCompletedHandler },
-                AsyncKind::OperationWithProgress => quote! { AsyncOperationWithProgressCompletedHandler },
+                AsyncKind::OperationWithProgress => {
+                    quote! { AsyncOperationWithProgressCompletedHandler }
+                }
                 _ => unimplemented!(),
             };
 
@@ -618,13 +672,28 @@ impl<'a> Gen<'a> {
             }
         }
     }
-    pub fn interface_winrt_trait(&self, def: TypeDef, generics: &[Type], ident: &TokenStream, constraints: &TokenStream, _phantoms: &TokenStream, features: &TokenStream) -> TokenStream {
-        if self.reader.type_def_flags(def).contains(TypeAttributes::WINRT) {
+    pub fn interface_winrt_trait(
+        &self,
+        def: TypeDef,
+        generics: &[Type],
+        ident: &TokenStream,
+        constraints: &TokenStream,
+        _phantoms: &TokenStream,
+        features: &TokenStream,
+    ) -> TokenStream {
+        if self
+            .reader
+            .type_def_flags(def)
+            .contains(TypeAttributes::WINRT)
+        {
             let type_signature = if self.reader.type_def_kind(def) == TypeKind::Class {
-                let type_signature = Literal::byte_string(self.reader.type_def_signature(def, generics).as_bytes());
+                let type_signature =
+                    Literal::byte_string(self.reader.type_def_signature(def, generics).as_bytes());
                 quote! { ::windows::core::ConstBuffer::from_slice(#type_signature) }
             } else {
-                let signature = Literal::byte_string(format!("{{{:#?}}}", self.reader.type_def_guid(def).unwrap()).as_bytes());
+                let signature = Literal::byte_string(
+                    format!("{{{:#?}}}", self.reader.type_def_guid(def).unwrap()).as_bytes(),
+                );
 
                 if generics.is_empty() {
                     quote! { ::windows::core::ConstBuffer::from_slice(#signature) }
@@ -672,8 +741,19 @@ impl<'a> Gen<'a> {
             quote! {}
         }
     }
-    pub fn runtime_name_trait(&self, def: TypeDef, _generics: &[Type], name: &TokenStream, constraints: &TokenStream, features: &TokenStream) -> TokenStream {
-        if self.reader.type_def_flags(def).contains(TypeAttributes::WINRT) {
+    pub fn runtime_name_trait(
+        &self,
+        def: TypeDef,
+        _generics: &[Type],
+        name: &TokenStream,
+        constraints: &TokenStream,
+        features: &TokenStream,
+    ) -> TokenStream {
+        if self
+            .reader
+            .type_def_flags(def)
+            .contains(TypeAttributes::WINRT)
+        {
             // TODO: this needs to use a ConstBuffer-like facility to accomodate generics
             let runtime_name = format!("{}", self.reader.type_def_type_name(def));
 
@@ -691,7 +771,15 @@ impl<'a> Gen<'a> {
         }
     }
 
-    pub fn interface_trait(&self, def: TypeDef, generics: &[Type], ident: &TokenStream, constraints: &TokenStream, features: &TokenStream, has_unknown_base: bool) -> TokenStream {
+    pub fn interface_trait(
+        &self,
+        def: TypeDef,
+        generics: &[Type],
+        ident: &TokenStream,
+        constraints: &TokenStream,
+        features: &TokenStream,
+        has_unknown_base: bool,
+    ) -> TokenStream {
         if let Some(default) = self.reader.type_def_default_interface(def) {
             let default_name = self.type_name(&default);
             let vtbl = self.type_vtbl_name(&default);
@@ -741,7 +829,14 @@ impl<'a> Gen<'a> {
             tokens
         }
     }
-    pub fn interface_vtbl(&self, def: TypeDef, generics: &[Type], _ident: &TokenStream, constraints: &TokenStream, features: &TokenStream) -> TokenStream {
+    pub fn interface_vtbl(
+        &self,
+        def: TypeDef,
+        generics: &[Type],
+        _ident: &TokenStream,
+        constraints: &TokenStream,
+        features: &TokenStream,
+    ) -> TokenStream {
         let vtbl = self.type_def_vtbl_name(def, generics);
         let mut methods = quote! {};
         let mut method_names = MethodNames::new();
@@ -749,8 +844,12 @@ impl<'a> Gen<'a> {
         let phantoms = self.generic_named_phantoms(generics);
 
         match self.reader.type_def_vtables(def).last() {
-            Some(Type::IUnknown) => methods.combine(&quote! { pub base__: ::windows::core::IUnknown_Vtbl, }),
-            Some(Type::IInspectable) => methods.combine(&quote! { pub base__: ::windows::core::IInspectable_Vtbl, }),
+            Some(Type::IUnknown) => {
+                methods.combine(&quote! { pub base__: ::windows::core::IUnknown_Vtbl, })
+            }
+            Some(Type::IInspectable) => {
+                methods.combine(&quote! { pub base__: ::windows::core::IInspectable_Vtbl, })
+            }
             Some(Type::TypeDef((def, _))) => {
                 let vtbl = self.type_def_vtbl_name(*def, &[]);
                 methods.combine(&quote! { pub base__: #vtbl, });
@@ -792,18 +891,34 @@ impl<'a> Gen<'a> {
             }
         }
     }
-    pub fn vtbl_signature(&self, def: TypeDef, _generics: &[Type], signature: &Signature) -> TokenStream {
-        let is_winrt = self.reader.type_def_flags(def).contains(TypeAttributes::WINRT);
+    pub fn vtbl_signature(
+        &self,
+        def: TypeDef,
+        _generics: &[Type],
+        signature: &Signature,
+    ) -> TokenStream {
+        let is_winrt = self
+            .reader
+            .type_def_flags(def)
+            .contains(TypeAttributes::WINRT);
         let hresult = self.type_name(&Type::HRESULT);
 
         let (trailing_return_type, return_type, udt_return_type) = if is_winrt {
             if let Some(return_type) = &signature.return_type {
                 if let Type::WinrtArray(kind) = return_type {
                     let tokens = self.type_abi_name(kind);
-                    (quote! { result_size__: *mut u32, result__: *mut *mut #tokens }, quote! { -> #hresult }, quote! {})
+                    (
+                        quote! { result_size__: *mut u32, result__: *mut *mut #tokens },
+                        quote! { -> #hresult },
+                        quote! {},
+                    )
                 } else {
                     let tokens = self.type_abi_name(return_type);
-                    (quote! { result__: *mut #tokens }, quote! { -> #hresult }, quote! {})
+                    (
+                        quote! { result__: *mut #tokens },
+                        quote! { -> #hresult },
+                        quote! {},
+                    )
                 }
             } else {
                 (quote! {}, quote! { -> #hresult }, quote! {})
@@ -824,9 +939,14 @@ impl<'a> Gen<'a> {
             let name = self.param_name(p.def);
             if is_winrt {
                 let abi = self.type_abi_name(&p.ty);
-                let abi_size_name: TokenStream = format!("{}_array_size", self.reader.param_name(p.def)).into();
+                let abi_size_name: TokenStream =
+                    format!("{}_array_size", self.reader.param_name(p.def)).into();
 
-                if self.reader.param_flags(p.def).contains(ParamAttributes::INPUT) {
+                if self
+                    .reader
+                    .param_flags(p.def)
+                    .contains(ParamAttributes::INPUT)
+                {
                     if p.ty.is_winrt_array() {
                         quote! { #abi_size_name: u32, #name: *const #abi, }
                     } else if p.ty.is_winrt_const_ref() {
@@ -880,20 +1000,26 @@ impl<'a> Gen<'a> {
                 SignatureKind::Query(query) if query.object == position => {
                     quote! { result__.as_mut_ptr(), }
                 }
-                SignatureKind::ReturnValue | SignatureKind::ResultValue if params.len() - 1 == position => {
+                SignatureKind::ReturnValue | SignatureKind::ResultValue
+                    if params.len() - 1 == position =>
+                {
                     quote! { result__.as_mut_ptr(), }
                 }
                 SignatureKind::QueryOptional(query) if query.object == position => {
                     quote! { result__ as *mut _ as *mut _, }
                 }
-                SignatureKind::Query(query) | SignatureKind::QueryOptional(query) if query.guid == position => {
+                SignatureKind::Query(query) | SignatureKind::QueryOptional(query)
+                    if query.guid == position =>
+                {
                     quote! { &<T as ::windows::core::Interface>::IID, }
                 }
                 _ => {
                     let name = self.param_name(param.def);
                     let flags = self.reader.param_flags(param.def);
                     match param.kind {
-                        SignatureParamKind::ArrayFixed(_) | SignatureParamKind::ArrayRelativeLen(_) | SignatureParamKind::ArrayRelativeByteLen(_) => {
+                        SignatureParamKind::ArrayFixed(_)
+                        | SignatureParamKind::ArrayRelativeLen(_)
+                        | SignatureParamKind::ArrayRelativeByteLen(_) => {
                             let map = if flags.contains(ParamAttributes::OPTIONAL) {
                                 quote! { #name.as_deref().map_or(::core::ptr::null(), |slice|slice.as_ptr()) }
                             } else {
@@ -954,7 +1080,9 @@ impl<'a> Gen<'a> {
                         continue;
                     }
                 }
-                SignatureKind::ReturnValue | SignatureKind::ResultValue if params.len() - 1 == position => {
+                SignatureKind::ReturnValue | SignatureKind::ResultValue
+                    if params.len() - 1 == position =>
+                {
                     continue;
                 }
                 _ => {}
@@ -967,12 +1095,20 @@ impl<'a> Gen<'a> {
                     let ty = param.ty.deref();
                     let ty = self.type_default_name(&ty);
                     let len = Literal::u32_unsuffixed(fixed as _);
-                    let ty = if self.reader.param_flags(param.def).contains(ParamAttributes::OUTPUT) {
+                    let ty = if self
+                        .reader
+                        .param_flags(param.def)
+                        .contains(ParamAttributes::OUTPUT)
+                    {
                         quote! { &mut [#ty; #len] }
                     } else {
                         quote! { &[#ty; #len] }
                     };
-                    if self.reader.param_flags(param.def).contains(ParamAttributes::OPTIONAL) {
+                    if self
+                        .reader
+                        .param_flags(param.def)
+                        .contains(ParamAttributes::OPTIONAL)
+                    {
                         tokens.combine(&quote! { #name: ::core::option::Option<#ty>, });
                     } else {
                         tokens.combine(&quote! { #name: #ty, });
@@ -981,31 +1117,49 @@ impl<'a> Gen<'a> {
                 SignatureParamKind::ArrayRelativeLen(_) => {
                     let ty = param.ty.deref();
                     let ty = self.type_default_name(&ty);
-                    let ty = if self.reader.param_flags(param.def).contains(ParamAttributes::OUTPUT) {
+                    let ty = if self
+                        .reader
+                        .param_flags(param.def)
+                        .contains(ParamAttributes::OUTPUT)
+                    {
                         quote! { &mut [#ty] }
                     } else {
                         quote! { &[#ty] }
                     };
-                    if self.reader.param_flags(param.def).contains(ParamAttributes::OPTIONAL) {
+                    if self
+                        .reader
+                        .param_flags(param.def)
+                        .contains(ParamAttributes::OPTIONAL)
+                    {
                         tokens.combine(&quote! { #name: ::core::option::Option<#ty>, });
                     } else {
                         tokens.combine(&quote! { #name: #ty, });
                     }
                 }
                 SignatureParamKind::ArrayRelativeByteLen(_) => {
-                    let ty = if self.reader.param_flags(param.def).contains(ParamAttributes::OUTPUT) {
+                    let ty = if self
+                        .reader
+                        .param_flags(param.def)
+                        .contains(ParamAttributes::OUTPUT)
+                    {
                         quote! { &mut [u8] }
                     } else {
                         quote! { &[u8] }
                     };
-                    if self.reader.param_flags(param.def).contains(ParamAttributes::OPTIONAL) {
+                    if self
+                        .reader
+                        .param_flags(param.def)
+                        .contains(ParamAttributes::OPTIONAL)
+                    {
                         tokens.combine(&quote! { #name: ::core::option::Option<#ty>, });
                     } else {
                         tokens.combine(&quote! { #name: #ty, });
                     }
                 }
                 SignatureParamKind::ArrayRelativePtr(_) => {}
-                SignatureParamKind::TryInto | SignatureParamKind::IntoParam | SignatureParamKind::Into => {
+                SignatureParamKind::TryInto
+                | SignatureParamKind::IntoParam
+                | SignatureParamKind::Into => {
                     let (position, _) = generic_params.next().unwrap();
                     let kind: TokenStream = format!("P{position}").into();
                     tokens.combine(&quote! { #name: #kind, });
@@ -1029,9 +1183,16 @@ impl<'a> Gen<'a> {
     }
 
     pub fn impl_signature(&self, def: TypeDef, signature: &Signature) -> TokenStream {
-        if self.reader.type_def_flags(def).contains(TypeAttributes::WINRT) {
+        if self
+            .reader
+            .type_def_flags(def)
+            .contains(TypeAttributes::WINRT)
+        {
             let is_delegate = self.reader.type_def_kind(def) == TypeKind::Delegate;
-            let params = signature.params.iter().map(|p| self.winrt_produce_type(p, !is_delegate));
+            let params = signature
+                .params
+                .iter()
+                .map(|p| self.winrt_produce_type(p, !is_delegate));
 
             let return_type = if let Some(return_type) = &signature.return_type {
                 let tokens = self.type_name(return_type);
@@ -1068,7 +1229,9 @@ impl<'a> Gen<'a> {
 
             let return_type = match signature_kind {
                 SignatureKind::ReturnVoid => quote! {},
-                SignatureKind::Query(_) | SignatureKind::QueryOptional(_) | SignatureKind::ResultVoid => quote! { -> ::windows::core::Result<()> },
+                SignatureKind::Query(_)
+                | SignatureKind::QueryOptional(_)
+                | SignatureKind::ResultVoid => quote! { -> ::windows::core::Result<()> },
                 SignatureKind::ResultValue => {
                     let return_type = signature.params[signature.params.len() - 1].ty.deref();
                     let return_type = self.type_name(&return_type);
@@ -1084,7 +1247,11 @@ impl<'a> Gen<'a> {
     fn winrt_produce_type(&self, param: &SignatureParam, include_param_names: bool) -> TokenStream {
         let default_type = self.type_default_name(&param.ty);
 
-        let sig = if self.reader.param_flags(param.def).contains(ParamAttributes::INPUT) {
+        let sig = if self
+            .reader
+            .param_flags(param.def)
+            .contains(ParamAttributes::INPUT)
+        {
             if param.ty.is_winrt_array() {
                 quote! { &[#default_type] }
             } else if self.reader.type_is_primitive(&param.ty) {
@@ -1115,7 +1282,11 @@ impl<'a> Gen<'a> {
         let name = self.param_name(param.def);
         let kind = self.type_default_name(&param.ty);
 
-        if self.reader.param_flags(param.def).contains(ParamAttributes::INPUT) {
+        if self
+            .reader
+            .param_flags(param.def)
+            .contains(ParamAttributes::INPUT)
+        {
             if self.reader.type_is_primitive(&param.ty) {
                 quote! { #name: #kind, }
             } else if self.reader.type_is_nullable(&param.ty) {
@@ -1133,7 +1304,12 @@ impl<'a> Gen<'a> {
 pub fn to_ident(name: &str) -> TokenStream {
     // keywords list based on https://doc.rust-lang.org/reference/keywords.html
     match name {
-        "abstract" | "as" | "become" | "box" | "break" | "const" | "continue" | "crate" | "do" | "else" | "enum" | "extern" | "false" | "final" | "fn" | "for" | "if" | "impl" | "in" | "let" | "loop" | "macro" | "match" | "mod" | "move" | "mut" | "override" | "priv" | "pub" | "ref" | "return" | "static" | "struct" | "super" | "trait" | "true" | "type" | "typeof" | "unsafe" | "unsized" | "use" | "virtual" | "where" | "while" | "yield" | "try" | "async" | "await" | "dyn" => format!("r#{name}").into(),
+        "abstract" | "as" | "become" | "box" | "break" | "const" | "continue" | "crate" | "do"
+        | "else" | "enum" | "extern" | "false" | "final" | "fn" | "for" | "if" | "impl" | "in"
+        | "let" | "loop" | "macro" | "match" | "mod" | "move" | "mut" | "override" | "priv"
+        | "pub" | "ref" | "return" | "static" | "struct" | "super" | "trait" | "true" | "type"
+        | "typeof" | "unsafe" | "unsized" | "use" | "virtual" | "where" | "while" | "yield"
+        | "try" | "async" | "await" | "dyn" => format!("r#{name}").into(),
         "Self" | "self" => format!("{name}_").into(),
         "_" => "unused".into(),
         _ => trim_tick(name).into(),
@@ -1207,9 +1383,21 @@ mod tests {
 
     #[test]
     fn test_starts_with() {
-        assert!(starts_with("Windows.Win32.Graphics.Direct3D11on12", "Windows.Win32.Graphics.Direct3D11on12"));
-        assert!(starts_with("Windows.Win32.Graphics.Direct3D11on12", "Windows.Win32.Graphics"));
-        assert!(!starts_with("Windows.Win32.Graphics.Direct3D11on12", "Windows.Win32.Graphics.Direct3D11"));
-        assert!(!starts_with("Windows.Win32.Graphics.Direct3D", "Windows.Win32.Graphics.Direct3D11"));
+        assert!(starts_with(
+            "Windows.Win32.Graphics.Direct3D11on12",
+            "Windows.Win32.Graphics.Direct3D11on12"
+        ));
+        assert!(starts_with(
+            "Windows.Win32.Graphics.Direct3D11on12",
+            "Windows.Win32.Graphics"
+        ));
+        assert!(!starts_with(
+            "Windows.Win32.Graphics.Direct3D11on12",
+            "Windows.Win32.Graphics.Direct3D11"
+        ));
+        assert!(!starts_with(
+            "Windows.Win32.Graphics.Direct3D",
+            "Windows.Win32.Graphics.Direct3D11"
+        ));
     }
 }
