+++ crates/libs/bindgen/src/structs.rs
@@ -38,11 +38,17 @@ fn gen_struct_with_name(gen: &Gen, def: TypeDef, struct_name: &str, cfg: &Cfg) -
 
         if gen.reader.field_flags(f).contains(FieldAttributes::LITERAL) {
             quote! {}
-        } else if !gen.sys && flags.contains(TypeAttributes::EXPLICIT_LAYOUT) && !gen.reader.field_is_copyable(f, def) {
+        } else if !gen.sys
+            && flags.contains(TypeAttributes::EXPLICIT_LAYOUT)
+            && !gen.reader.field_is_copyable(f, def)
+        {
             // Rust can't tell that the type is copyable and won't accept windows::core::ManuallyDrop
             let ty = gen.type_default_name(&ty);
             quote! { pub #name: ::std::mem::ManuallyDrop<#ty>, }
-        } else if !gen.sys && !flags.contains(TypeAttributes::WINRT) && !gen.reader.field_is_blittable(f, def) {
+        } else if !gen.sys
+            && !flags.contains(TypeAttributes::WINRT)
+            && !gen.reader.field_is_blittable(f, def)
+        {
             if let Type::Win32Array((ty, len)) = ty {
                 let ty = gen.type_name(&ty);
                 quote! { pub #name: [::windows::core::ManuallyDrop<#ty>; #len], }
@@ -87,8 +93,6 @@ fn gen_struct_with_name(gen: &Gen, def: TypeDef, struct_name: &str, cfg: &Cfg) -
                 }
             }
         });
-
-        tokens.combine(&extensions::gen(gen.reader.type_def_type_name(def)));
     }
 
     for (index, nested_type) in gen.reader.nested_types(def).enumerate() {
@@ -118,8 +122,13 @@ fn gen_windows_traits(gen: &Gen, def: TypeDef, name: &TokenStream, cfg: &Cfg) ->
             }
         };
 
-        if gen.reader.type_def_flags(def).contains(TypeAttributes::WINRT) {
-            let signature = Literal::byte_string(gen.reader.type_def_signature(def, &[]).as_bytes());
+        if gen
+            .reader
+            .type_def_flags(def)
+            .contains(TypeAttributes::WINRT)
+        {
+            let signature =
+                Literal::byte_string(gen.reader.type_def_signature(def, &[]).as_bytes());
 
             let clone = if gen.reader.type_def_is_blittable(def) {
                 quote! { *from }
@@ -146,7 +155,11 @@ fn gen_windows_traits(gen: &Gen, def: TypeDef, name: &TokenStream, cfg: &Cfg) ->
 fn gen_compare_traits(gen: &Gen, def: TypeDef, name: &TokenStream, cfg: &Cfg) -> TokenStream {
     let features = gen.cfg_features(cfg);
 
-    if gen.sys || gen.reader.type_def_has_explicit_layout(def) || gen.reader.type_def_has_packing(def) || gen.reader.type_def_has_callback(def) {
+    if gen.sys
+        || gen.reader.type_def_has_explicit_layout(def)
+        || gen.reader.type_def_has_packing(def)
+        || gen.reader.type_def_has_callback(def)
+    {
         quote! {}
     } else {
         let fields = gen.reader.type_def_fields(def).filter_map(|f| {
@@ -172,7 +185,10 @@ fn gen_compare_traits(gen: &Gen, def: TypeDef, name: &TokenStream, cfg: &Cfg) ->
 }
 
 fn gen_debug(gen: &Gen, def: TypeDef, ident: &TokenStream, cfg: &Cfg) -> TokenStream {
-    if gen.sys || gen.reader.type_def_has_explicit_layout(def) || gen.reader.type_def_has_packing(def) {
+    if gen.sys
+        || gen.reader.type_def_has_explicit_layout(def)
+        || gen.reader.type_def_has_packing(def)
+    {
         quote! {}
     } else {
         let name = ident.as_str();
@@ -221,7 +237,11 @@ fn gen_copy_clone(gen: &Gen, def: TypeDef, name: &TokenStream, cfg: &Cfg) -> Tok
     } else if gen.reader.type_def_class_layout(def).is_some() {
         // Don't support copy/clone of packed structs: https://github.com/rust-lang/rust/issues/82523
         quote! {}
-    } else if !gen.reader.type_def_flags(def).contains(TypeAttributes::WINRT) {
+    } else if !gen
+        .reader
+        .type_def_flags(def)
+        .contains(TypeAttributes::WINRT)
+    {
         quote! {
             #features
             impl ::core::clone::Clone for #name {
@@ -253,7 +273,12 @@ fn gen_copy_clone(gen: &Gen, def: TypeDef, name: &TokenStream, cfg: &Cfg) -> Tok
     }
 }
 
-fn gen_struct_constants(gen: &Gen, def: TypeDef, struct_name: &TokenStream, cfg: &Cfg) -> TokenStream {
+fn gen_struct_constants(
+    gen: &Gen,
+    def: TypeDef,
+    struct_name: &TokenStream,
+    cfg: &Cfg,
+) -> TokenStream {
     let features = gen.cfg_features(cfg);
 
     let constants = gen.reader.type_def_fields(def).filter_map(|f| {
