+++ crates/libs/windows/src/Windows/Win32/Networking/WinSock/mod.rs
@@ -8811,16 +8811,6 @@ impl ::core::default::Default for IN6_ADDR {
         unsafe { ::core::mem::zeroed() }
     }
 }
-impl ::core::convert::From<::std::net::Ipv6Addr> for IN6_ADDR {
-    fn from(addr: ::std::net::Ipv6Addr) -> Self {
-        Self { u: IN6_ADDR_0 { Byte: addr.octets() } }
-    }
-}
-impl ::core::convert::From<IN6_ADDR> for ::std::net::Ipv6Addr {
-    fn from(in6_addr: IN6_ADDR) -> Self {
-        Self::from(unsafe { in6_addr.u.Byte })
-    }
-}
 #[repr(C)]
 #[doc = "*Required features: `\"Win32_Networking_WinSock\"`*"]
 pub union IN6_ADDR_0 {
@@ -9088,16 +9078,6 @@ unsafe impl ::windows::core::Abi for IN_ADDR {
 impl ::core::default::Default for IN_ADDR {
     fn default() -> Self {
         unsafe { ::core::mem::zeroed() }
-    }
-}
-impl ::core::convert::From<::std::net::Ipv4Addr> for IN_ADDR {
-    fn from(addr: ::std::net::Ipv4Addr) -> Self {
-        Self { S_un: IN_ADDR_0 { S_addr: u32::from(addr).to_be() } }
-    }
-}
-impl ::core::convert::From<IN_ADDR> for ::std::net::Ipv4Addr {
-    fn from(in_addr: IN_ADDR) -> Self {
-        Self::from(u32::from_be(unsafe { in_addr.S_un.S_addr }))
     }
 }
 #[repr(C)]
@@ -13424,11 +13404,6 @@ impl ::core::default::Default for SOCKADDR_IN {
         unsafe { ::core::mem::zeroed() }
     }
 }
-impl ::core::convert::From<::std::net::SocketAddrV4> for SOCKADDR_IN {
-    fn from(addr: ::std::net::SocketAddrV4) -> Self {
-        SOCKADDR_IN { sin_family: AF_INET, sin_port: addr.port().to_be(), sin_addr: (*addr.ip()).into(), ..Default::default() }
-    }
-}
 #[repr(C)]
 #[doc = "*Required features: `\"Win32_Networking_WinSock\"`*"]
 pub struct SOCKADDR_IN6 {
@@ -13450,18 +13425,6 @@ unsafe impl ::windows::core::Abi for SOCKADDR_IN6 {
 impl ::core::default::Default for SOCKADDR_IN6 {
     fn default() -> Self {
         unsafe { ::core::mem::zeroed() }
-    }
-}
-impl ::core::convert::From<::std::net::SocketAddrV6> for SOCKADDR_IN6 {
-    fn from(addr: ::std::net::SocketAddrV6) -> Self {
-        SOCKADDR_IN6 {
-            sin6_family: AF_INET6,
-            sin6_port: addr.port().to_be(),
-            sin6_flowinfo: addr.flowinfo().to_be(),
-            sin6_addr: (*addr.ip()).into(),
-            Anonymous: SOCKADDR_IN6_0 { sin6_scope_id: addr.scope_id() },
-            ..Default::default()
-        }
     }
 }
 #[repr(C)]
@@ -13562,24 +13525,6 @@ unsafe impl ::windows::core::Abi for SOCKADDR_INET {
 impl ::core::default::Default for SOCKADDR_INET {
     fn default() -> Self {
         unsafe { ::core::mem::zeroed() }
-    }
-}
-impl ::core::convert::From<::std::net::SocketAddrV4> for SOCKADDR_INET {
-    fn from(addr: ::std::net::SocketAddrV4) -> Self {
-        SOCKADDR_INET { Ipv4: addr.into() }
-    }
-}
-impl ::core::convert::From<::std::net::SocketAddrV6> for SOCKADDR_INET {
-    fn from(addr: ::std::net::SocketAddrV6) -> Self {
-        SOCKADDR_INET { Ipv6: addr.into() }
-    }
-}
-impl ::core::convert::From<::std::net::SocketAddr> for SOCKADDR_INET {
-    fn from(addr: ::std::net::SocketAddr) -> Self {
-        match addr {
-            ::std::net::SocketAddr::V4(socket_addr_v4) => socket_addr_v4.into(),
-            ::std::net::SocketAddr::V6(socket_addr_v6) => socket_addr_v6.into(),
-        }
     }
 }
 #[repr(C)]
@@ -17513,5 +17458,72 @@ pub type LPWSPSTARTUP = ::core::option::Option<unsafe extern "system" fn(wversio
 #[doc = "*Required features: `\"Win32_Networking_WinSock\"`, `\"Win32_Foundation\"`*"]
 #[cfg(feature = "Win32_Foundation")]
 pub type LPWSPSTRINGTOADDRESS = ::core::option::Option<unsafe extern "system" fn(addressstring: ::windows::core::PCWSTR, addressfamily: i32, lpprotocolinfo: *const WSAPROTOCOL_INFOW, lpaddress: *mut SOCKADDR, lpaddresslength: *mut i32, lperrno: *mut i32) -> i32>;
+impl ::core::convert::From<::std::net::Ipv4Addr> for IN_ADDR {
+    fn from(addr: ::std::net::Ipv4Addr) -> Self {
+        // u32::from(addr) is in host byte order
+        // S_addr must be big-endian, network byte order
+        Self { S_un: IN_ADDR_0 { S_addr: u32::from(addr).to_be() } }
+    }
+}
+impl ::core::convert::From<IN_ADDR> for ::std::net::Ipv4Addr {
+    fn from(in_addr: IN_ADDR) -> Self {
+        // SAFETY: this is safe because the union variants are just views of the same exact data
+        // in_addr.S_un.S_addr is big-endian, network byte order
+        // Ipv4Addr::new() expects the parameter in host byte order
+        Self::from(u32::from_be(unsafe { in_addr.S_un.S_addr }))
+    }
+}
+impl ::core::convert::From<::std::net::Ipv6Addr> for IN6_ADDR {
+    fn from(addr: ::std::net::Ipv6Addr) -> Self {
+        Self { u: IN6_ADDR_0 { Byte: addr.octets() } }
+    }
+}
+impl ::core::convert::From<IN6_ADDR> for ::std::net::Ipv6Addr {
+    fn from(in6_addr: IN6_ADDR) -> Self {
+        // SAFETY: this is safe because the union variants are just views of the same exact data
+        Self::from(unsafe { in6_addr.u.Byte })
+    }
+}
+impl ::core::convert::From<::std::net::SocketAddrV4> for SOCKADDR_IN {
+    fn from(addr: ::std::net::SocketAddrV4) -> Self {
+        // addr.port() is in host byte order
+        // sin_port must be big-endian, network byte order
+        SOCKADDR_IN { sin_family: AF_INET, sin_port: addr.port().to_be(), sin_addr: (*addr.ip()).into(), ..Default::default() }
+    }
+}
+impl ::core::convert::From<::std::net::SocketAddrV6> for SOCKADDR_IN6 {
+    fn from(addr: ::std::net::SocketAddrV6) -> Self {
+        // addr.port() and addr.flowinfo() are in host byte order
+        // sin6_port and sin6_flowinfo must be big-endian, network byte order
+        // sin6_scope_id is a bitfield without endianness
+        SOCKADDR_IN6 {
+            sin6_family: AF_INET6,
+            sin6_port: addr.port().to_be(),
+            sin6_flowinfo: addr.flowinfo().to_be(),
+            sin6_addr: (*addr.ip()).into(),
+            Anonymous: SOCKADDR_IN6_0 { sin6_scope_id: addr.scope_id() },
+            ..Default::default()
+        }
+    }
+}
+
+impl ::core::convert::From<::std::net::SocketAddrV4> for SOCKADDR_INET {
+    fn from(addr: ::std::net::SocketAddrV4) -> Self {
+        SOCKADDR_INET { Ipv4: addr.into() }
+    }
+}
+impl ::core::convert::From<::std::net::SocketAddrV6> for SOCKADDR_INET {
+    fn from(addr: ::std::net::SocketAddrV6) -> Self {
+        SOCKADDR_INET { Ipv6: addr.into() }
+    }
+}
+impl ::core::convert::From<::std::net::SocketAddr> for SOCKADDR_INET {
+    fn from(addr: ::std::net::SocketAddr) -> Self {
+        match addr {
+            ::std::net::SocketAddr::V4(socket_addr_v4) => socket_addr_v4.into(),
+            ::std::net::SocketAddr::V6(socket_addr_v6) => socket_addr_v6.into(),
+        }
+    }
+}
 #[cfg(feature = "implement")]
 ::core::include!("impl.rs");
